# Conflicts:
#	.gradle/8.5/executionHistory/executionHistory.bin
#	.gradle/8.5/executionHistory/executionHistory.lock
#	.gradle/8.5/fileHashes/fileHashes.bin
#	.gradle/8.5/fileHashes/fileHashes.lock
#	.gradle/8.5/fileHashes/resourceHashesCache.bin
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	.gradle/file-system.probe
#	.idea/shelf/Uncommitted_changes_before_Checkout_at_1_10_2024_12_30_PM__Changes_.xml
#	.idea/shelf/Uncommitted_changes_before_Update_at_1_10_2024_12_31_PM__Changes_.xml
#	.idea/workspace.xml
#	build/classes/java/main/com/example/triviaApplication/controllers/QuizController.class
#	build/classes/java/main/com/example/triviaApplication/helpers/QuizService.class
#	build/classes/java/main/com/example/triviaApplication/models/Quiz.class
#	build/tmp/compileJava/previous-compilation-data.bin
#	client/src/components/CreateQuizForm.js
#	client/src/components/QuestionForm.js
#	client/src/components/QuizPage.js
#	src/main/java/com/example/triviaApplication/Validator/QuizValidator.java
#	src/main/java/com/example/triviaApplication/controllers/QuizController.java
#	src/main/java/com/example/triviaApplication/helpers/QuizService.java
#	src/main/java/com/example/triviaApplication/models/Quiz.java

+++ client/src/components/QuestionForm.js
@@ -1,6 +1,7 @@
-import React, { useState, useEffect } from 'react';
-import { useParams, useNavigate } from 'react-router-dom';
+import React, {useEffect, useState} from 'react';
 import './QuestionForm.css';
+import axios from "axios";
+import {useNavigate, useParams} from "react-router-dom";
 
 function QuestionForm() {
     const [username, setUsername] = useState('');
@@ -9,46 +10,11 @@ function QuestionForm() {
     const [correctAnswerIndex, setCorrectAnswerIndex] = useState(-1);
     const [userQuestions, setUserQuestions] = useState([]);
     const [searchUsername, setSearchUsername] = useState('');
+    const {quizId} = useParams();
     const [keyword, setKeyword] = useState('');
     const [editMode, setEditMode] = useState(false);
     const [editQuestionId, setEditQuestionId] = useState(null);
     const [keywordSearchUsername, setKeywordSearchUsername] = useState('');
-
-    const { quizId } = useParams();
-    const navigate = useNavigate();
-
-    useEffect(() => {
-        if (quizId) {
-            fetchQuizDetails();
-        }
-    }, [quizId]);
-
-    const fetchQuizDetails = async () => {
-        try {
-            const quizResponse = await fetch(`http://localhost:8080/quiz/${quizId}`);
-            if (!quizResponse.ok) throw new Error(`HTTP error! Status: ${quizResponse.status}`);
-            const quizData = await quizResponse.json();
-
-            if (quizData && quizData.user && quizData.user.id) {
-                fetchUserQuestionsForQuiz(quizData.user.id);
-            } else {
-                console.error('User data not found in quiz details');
-            }
-        } catch (error) {
-            console.error('Error fetching quiz details:', error);
-        }
-    };
-
-    const fetchUserQuestionsForQuiz = async (userId) => {
-        try {
-            const response = await fetch(`http://localhost:8080/users/${userId}/created-questions`);
-            if (!response.ok) throw new Error(`HTTP error! Status: ${response.status}`);
-            const questions = await response.json();
-            setUserQuestions(questions);
-        } catch (error) {
-            console.error('Error fetching user questions:', error);
-        }
-    };
 
     const handleAnswerChange = (index, event) => {
         const newAnswers = answers.map((answer, i) => {
@@ -69,66 +35,55 @@ function QuestionForm() {
         setCorrectAnswerIndex(index);
     };
 
+    const handleSubmit = async (event) => {
+        event.preventDefault();
 
-const handleSubmit = async (event) => {
-    event.preventDefault();
-    const questionData = {
-        user: { username },
-        text: questionText,
-        answers: answers.map((answer, index) => ({ ...answer, isCorrect: index === correctAnswerIndex })),
+        if (!username || !questionText || answers.some(answer => !answer.text)) {
+            alert('Please fill out all fields.');
+            return;
+        }
+
+        //Iryna added validation for correct answer checked
+        if (correctAnswerIndex === -1) {
+            alert('Please select at least one correct answer.');
+            return;
+        }
+
+        const questionData = {
+            user: { username },
+            text: questionText,
+            answers
+        };
+
+        try {
+            //Iryna changed
+            const response =  axios.post('http://localhost:8080/question', questionData)
+                // Iryna added
+                .then(responseData => {
+                    if (responseData.data != null && responseData.data.id !=null){
+                        if (quizId !=null) {
+                            axios.post(`/quiz/addQuestion/${quizId}`, responseData.data.id.toString());
+                        }
+                    }
+                });
+            //Iryna end
+
+            alert('Question saved successfully!');
+            setUsername('');
+            setQuestionText('');
+            setAnswers(new Array(4).fill({ text: '', isCorrect: false }));
+            setCorrectAnswerIndex(-1);
+        } catch (error) {
+            console.error('There was an error saving the question:', error);
+            alert('Error saving question.');
+        }
     };
-
-    const method = editMode ? 'PUT' : 'POST';
-    const questionUrl = editMode ? `http://localhost:8080/question/${editQuestionId}` : 'http://localhost:8080/question';
-
-    try {
-        const response = await fetch(questionUrl, {
-            method: method,
-            headers: { 'Content-Type': 'application/json' },
-            body: JSON.stringify(questionData),
-        });
-        if (!response.ok) throw new Error(`HTTP error! Status: ${response.status}`);
-
-        const savedQuestion = await response.json();
-
-        // Check if quizId is defined and not in edit mode
-        if (!editMode && quizId && quizId !== 'undefined') {
-            const addQuestionToQuizResponse = await fetch(`http://localhost:8080/quiz/addQuestion/${quizId}`, {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ questionId: savedQuestion.id }),
-            });
-            if (!addQuestionToQuizResponse.ok) throw new Error(`HTTP error! Status: ${addQuestionToQuizResponse.status}`);
-        }
-
-        alert('Question saved successfully!');
-
-        // Reset form
-        setQuestionText('');
-        setAnswers(new Array(4).fill({ text: '', isCorrect: false }));
-        setCorrectAnswerIndex(-1);
-        setEditMode(false);
-        setEditQuestionId(null);
-
-        // Only navigate if quizId is valid
-        if (quizId && quizId !== 'undefined') {
-            navigate(`/question-form/${quizId}`);
-        } else {
-            navigate(`/question-form`);
-        }
-
-    } catch (error) {
-        alert('Error saving question.');
-        console.error('There was an error:', error);
-    }
-};
-
     const handleSearch = async () => {
         try {
             const response = await fetch(`http://localhost:8080/users/${searchUsername}/questions`);
-            if (!response.ok) throw new Error(`HTTP error! Status: ${response.status}`);
-            const questions = await response.json();
-            setUserQuestions(questions);
+            if (!response.ok) { throw new Error(`HTTP error! Status: ${response.status}`); }
+            const data = await response.json();
+            setUserQuestions(data);
         } catch (error) {
             console.error('Error fetching questions:', error);
             alert('Error fetching questions.');
@@ -153,16 +108,24 @@ const handleSubmit = async (event) => {
     };
 
     const handleDelete = async (questionId) => {
+        //Iryna added confirmation
+        const userConfirmed = window.confirm('Are you sure you want to delete this question from the quiz?')
+        if (!userConfirmed) {
+            return;
+        }
+        //Iryna end change
         try {
             const response = await fetch(`http://localhost:8080/question/${questionId}`, {
                 method: 'DELETE',
             });
-            if (!response.ok) throw new Error(`HTTP error! Status: ${response.status}`);
-            setUserQuestions(userQuestions.filter((question) => question.id !== questionId));
+
+            if (!response.ok) { throw new Error(`HTTP error! Status: ${response.status}`); }
+
+            setUserQuestions(prevQuestions => prevQuestions.filter(question => question.id !== questionId));
             alert('Question deleted successfully!');
         } catch (error) {
-            alert('Error deleting question.');
             console.error('Error deleting question:', error);
+            alert('Error deleting question.');
         }
     };
 
@@ -170,44 +133,25 @@ const handleSubmit = async (event) => {
         setUsername(question.user.username);
         setQuestionText(question.text);
         setAnswers(question.answers);
-        setCorrectAnswerIndex(question.answers.findIndex((answer) => answer.isCorrect));
+        const correctIndex = question.answers.findIndex(answer => answer.isCorrect);
+        setCorrectAnswerIndex(correctIndex);
         setEditMode(true);
         setEditQuestionId(question.id);
     };
 
-    const handleAddToQuiz = async (questionId) => {
-        try {
-             console.log(`quizId: ${quizId}`);
-             console.log(`questionId: ${questionId}`);
-             const requestBody = JSON.stringify({ questionId: questionId }); // Construct the request body as a JSON object
-             console.log(`quizId: ${quizId}`);
-             const response = await fetch(`http://localhost:8080/quiz/addQuestion/${quizId}`, {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: requestBody, // Provide the request body here
-            });
-            if (!response.ok) throw new Error(`HTTP error! Status: ${response.status}`);
-            alert('Question added to quiz successfully!');
-        } catch (error) {
-            alert('Error adding question to quiz.');
-            console.error('Error:', error);
-        }
-    };
-
-
     return (
-        <div>
-            <h2>Create a New Question</h2>
-            <form onSubmit={handleSubmit}>
-                <div>
-                    <label>Username:</label>
-                    <input type="text" value={username} onChange={(e) => setUsername(e.target.value)} />
-                </div>
-                <div>
-                    <label>Question:</label>
-                    <input type="text" value={questionText} onChange={(e) => setQuestionText(e.target.value)} />
-                </div>
-                {answers.map((answer, index) => (
+    <div>
+        <h2>Create a New Question</h2>
+        <form onSubmit={handleSubmit}>
+            <div>
+                <label>Username:</label>
+                <input type="text" value={username} onChange={(e) => setUsername(e.target.value)}/>
+            </div>
+            <div>
+                <label>Question:</label>
+                <input type="text" value={questionText} onChange={(e) => setQuestionText(e.target.value)}/>
+            </div>
+            {answers.map((answer, index) => (
                     <div key={index}>
                         <label>
                             Answer {index + 1}:
@@ -235,52 +179,51 @@ const handleSubmit = async (event) => {
                     placeholder="Username"
                 />
                 <button type="button" onClick={handleSearch}>Search</button>
-            </div>
 
-            <div>
-                <h2>Search by Keyword</h2>
-                <input
-                    type="text"
-                    value={keywordSearchUsername}
-                    onChange={(e) => setKeywordSearchUsername(e.target.value)}
-                    placeholder="Username for Keyword Search"
-                />
-                <input
-                    type="text"
-                    value={keyword}
-                    onChange={(e) => setKeyword(e.target.value)}
-                    placeholder="Keyword"
-                />
-                <button type="button" onClick={handleKeywordSearch}>Search Keyword</button>
-            </div>
+                <div>
+                    <h2>Search by Keyword</h2>
+                    <input
+                        type="text"
+                        value={keywordSearchUsername}
+                        onChange={(e) => setKeywordSearchUsername(e.target.value)}
+                        placeholder="Username for Keyword Search"
+                    />
+                    <input
+                        type="text"
+                        value={keyword}
+                        onChange={(e) => setKeyword(e.target.value)}
+                        placeholder="Keyword"
+                    />
+                    <button type="button" onClick={handleKeywordSearch}>Search Keyword</button>
+                </div>
 
-            <div>
-                <h3>Search Results</h3>
-                {userQuestions.map((question, index) => (
-                    <div key={index} className="question-item">
-                        <div className="question-content">
-                            <h3>Question {index + 1}</h3>
-                            <p>Question Text: {question.text}</p>
-                            <p>Question User: {question.user.username}</p>
-                            <h4>Answers:</h4>
-                            <ul>
-                                {question.answers.map((answer, ansIndex) => (
-                                    <li key={ansIndex}>
-                                        Answer {ansIndex + 1}: {answer.text}
-                                        {answer.isCorrect ? ' (Correct)' : ''}
-                                    </li>
-                                ))}
-                            </ul>
+                <div>
+                    <h3>Search Results</h3>
+                    {userQuestions.map((question, index) => (
+                        <div key={index} className="question-item">
+                            <div className="question-content">
+                                <h3>Question {index + 1}</h3>
+                                <p>Question Text: {question.text}</p>
+                                <p>Question User: {question.user.username}</p>
+                                <h4>Answers:</h4>
+                                <ul>
+                                    {question.answers.map((answer, ansIndex) => (
+                                        <li key={ansIndex}>
+                                            Answer {ansIndex + 1}: {answer.text}
+                                            {answer.isCorrect ? ' (Correct)' : ''}
+                                        </li>
+                                    ))}
+                                </ul>
+                            </div>
+                            <div className="question-actions">
+
+                                <button onClick={() => handleEdit(question)} className="question-button edit-button">Edit</button>
+
+                                <button onClick={() => handleDelete(question.id)} className="question-button delete-button">Delete</button>
+                            </div>
                         </div>
-                        <div className="question-actions">
-                            <button onClick={() => handleEdit(question)} className="question-button edit-button">Edit</button>
-                            <button onClick={() => handleDelete(question.id)} className="question-button delete-button">Delete</button>
-                            {quizId && (
-                                <button onClick={() => handleAddToQuiz(question.id)} className="question-button add-button">Add to Quiz</button>
-                            )}
-                        </div>
-                    </div>
-                ))}
+                    ))}
+                </div>
             </div>
         </div>
     );
